Refactoring script 'taxonomy_parser'
  - Moving function 'write_dfs' to scripts 'utils.py' module
